refactor: auth validation
