+++ src/components/Auth/FormikTextField/ValidationSpecialRules.types.ts
@@ -1,0 +1,6 @@
+export interface IValidationSpecialRules {
+  name: string;
+  min: number;
+  max: number;
+  countable: string;
+}
