+++ src/components/Auth/Login/Login.type.ts
@@ -1,0 +1,4 @@
+export interface ILogin {
+  email: string;
+  password: string;
+}
