+++ src/components/Auth/Registration/Registration.type.ts
@@ -1,0 +1,8 @@
+export interface IRegistration {
+  username: string,
+  age: string,
+  email: string,
+  password: string,
+  confirmPassword: string,
+  confirmEmail: string,
+}
